pyproject.toml
[project]
name = "meal-planner-agentic"
version = "0.1.0"
description = "Chat-driven weekly meal planner with Claude agents"
requires-python = ">=3.12"
dependencies = [
    "litellm>=1.52.0",
    "streamlit>=1.40.0",
    "pydantic>=2.8.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "cookidoo-api>=0.17.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "tools"]
py-modules = ["app", "models", "storage"]

[tool.pytest.ini_options]
testpaths = ["tests"]
